Move TransactionStatsPublisher to service package
